% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_check.R
\name{palette_check}
\alias{palette_check}
\title{Check that data matches the color palette}
\usage{
palette_check(names.dat, palette.name, vis.flag = FALSE)
}
\arguments{
\item{names.dat}{vector of data identifiers. For example, the column of taxa names}

\item{palette.name}{Name of the palette. See \code{make_colorvec()} for options}

\item{vis.flag}{If true, make a geom_tile object to visualize the colors to be used for the data.}
}
\value{
optionally, a ggplot object visualizing the color palette as applied to the data.
}
\description{
Check that data matches the color palette
}
\examples{
dat = data.frame(vals = rnorm(4), species = c("acremonium", "aspergillus", "candida", "vibrio"))
palette_check(dat$species, "wolfe2014")
# Now with a new taxa not in the palette (because it's common milkweed).
dat = data.frame(vals = rnorm(5), species = c("acremonium", "aspergillus", "candida", "vibrio", "Asclepias syriaca"))
palette_check(dat$species, "wolfe2014", vis.flag = TRUE)
}
