% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_vis.R
\name{palette_vis}
\alias{palette_vis}
\title{Visualize the colors for a given palette}
\usage{
palette_vis(palette.name)
}
\arguments{
\item{palette.name}{Name of the palette being used, defaulting to "wolfe2014". See help("make_colorvec") for details.}
}
\value{
A ggplot object showing color tiles and associated species/identification names.
}
\description{
Visualize the colors for a given palette
}
\examples{
palette_vis("wolfe2014")
}
