% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_vis_.R
\name{palette_vis_}
\alias{palette_vis_}
\title{Visualize the colors for a given palette, internal function}
\usage{
palette_vis_(color.vec)
}
\arguments{
\item{color.vec}{Named character vector: entries are colors, names are species/identifiers}
}
\value{
A ggplot object showing color tiles and associated species/identification names.
}
\description{
Takes a color vector (rather than palette name) to visualize with geom_tile. Used in \code{palette_vis()}, also useful for development/experimenting.
}
\examples{
palette_vis_(c(turtle = 'lightblue', rat = 'coral', ostrich = 'goldenrod3'))
}
